# openapi/openapi.json
{"openapi":"3.1.0","info":{"title":"Albion Market Data Platform Local API","version":"1.0.0"},"servers":[{"url":"http://127.0.0.1:8787"}],"paths":{"/healthz":{"get":{"operationId":"getHealth","parameters":[],"responses":{"200":{"description":"Response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/readyz":{"get":{"operationId":"getReadiness","parameters":[],"responses":{"200":{"description":"Response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReadinessResponse"}}}},"503":{"description":"Response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReadinessResponse"}}}}}}},"/api/v1/status":{"get":{"operationId":"getStatus","parameters":[],"responses":{"200":{"description":"Response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusResponse"}}}}}}},"/api/v1/markets":{"get":{"operationId":"listMarkets","parameters":[{"name":"includeDisabled","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"Response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MarketsResponse"}}}}}}},"/api/v1/prices":{"get":{"operationId":"getPrices","parameters":[{"name":"server","in":"query","required":false,"schema":{"type":"string","enum":["west","east","europe"]}},{"name":"itemIds","in":"query","required":false,"schema":{"type":"string","maxLength":4096}},{"name":"itemId","in":"query","required":false,"schema":{"type":"string","maxLength":256}},{"name":"marketKey","in":"query","required":false,"schema":{"type":"string","maxLength":128}},{"name":"locationId","in":"query","required":false,"schema":{"type":"string","maxLength":128}},{"name":"location","in":"query","required":false,"schema":{"type":"string","maxLength":256}},{"name":"quality","in":"query","required":true,"schema":{"type":"integer","minimum":1,"maximum":5}}],"responses":{"200":{"description":"Response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PricesResponse"}}}}}}},"/api/v1/history":{"get":{"operationId":"listHistory","parameters":[{"name":"server","in":"query","required":false,"schema":{"type":"string","enum":["west","east","europe"]}},{"name":"itemId","in":"query","required":false,"schema":{"type":"string","maxLength":256}},{"name":"marketKey","in":"query","required":false,"schema":{"type":"string","maxLength":128}},{"name":"locationId","in":"query","required":false,"schema":{"type":"string","maxLength":128}},{"name":"location","in":"query","required":false,"schema":{"type":"string","maxLength":256}},{"name":"quality","in":"query","required":false,"schema":{"type":"integer","minimum":1,"maximum":5}},{"name":"period","in":"query","required":false,"schema":{"type":"string","maxLength":64}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","minimum":1,"maximum":500,"default":100}}],"responses":{"200":{"description":"Response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HistoryResponse"}}}}}}},"/api/v1/orders":{"get":{"operationId":"listOrders","parameters":[{"name":"server","in":"query","required":false,"schema":{"type":"string","enum":["west","east","europe"]}},{"name":"itemId","in":"query","required":false,"schema":{"type":"string","maxLength":256}},{"name":"marketKey","in":"query","required":false,"schema":{"type":"string","maxLength":128}},{"name":"locationId","in":"query","required":false,"schema":{"type":"string","maxLength":128}},{"name":"location","in":"query","required":false,"schema":{"type":"string","maxLength":256}},{"name":"quality","in":"query","required":false,"schema":{"type":"integer","minimum":1,"maximum":5}},{"name":"auctionType","in":"query","required":false,"schema":{"type":"string","maxLength":64}},{"name":"side","in":"query","required":false,"schema":{"type":"string","enum":["buy","sell"]}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","minimum":1,"maximum":5000,"default":100}}],"responses":{"200":{"description":"Response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrdersResponse"}}}}}}}},"components":{"schemas":{"ErrorResponse":{"type":"object","properties":{"error":{"type":"string"}}},"HealthResponse":{"type":"object","properties":{"status":{"type":"string","const":"ok"},"service":{"type":"string"}}},"ReadinessChecks":{"type":"object","properties":{"catalog":{"type":"string","enum":["ok","unavailable"]},"repository":{"type":"string","enum":["ok","unavailable"]}}},"ReadinessResponse":{"type":"object","properties":{"status":{"type":"string","enum":["ready","not_ready"]},"service":{"type":"string"},"checks":{"$ref":"#/components/schemas/ReadinessChecks"}}},"ItemDimension":{"type":"object","properties":{"albionId":{"type":"integer","format":"int32"},"id":{"type":"string"},"name":{"type":"string"}}},"LocationDimension":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"marketKey":{"type":"string"}}},"MarketDefinition":{"type":"object","properties":{"key":{"type":"string"},"name":{"type":"string"},"type":{"type":"string","enum":["regular","black-market"]},"cityLocationId":{"type":"string"},"marketLocationId":{"anyOf":[{"type":"string"},{"type":"null"}]},"enabled":{"type":"boolean"}}},"QualityDimension":{"type":"object","properties":{"id":{"type":"integer","minimum":1,"maximum":5},"name":{"type":"string"}}},"NormalizedHistoryPoint":{"type":"object","properties":{"timestamp":{"type":"string","format":"date-time"},"itemCount":{"type":"integer","format":"int64"},"totalSilver":{"type":"integer","format":"int64"},"averageUnitPrice":{"type":"number","format":"double"}}},"NormalizedHistorySummary":{"type":"object","properties":{"soldUnits":{"type":"integer","format":"int64"},"activeBuckets":{"type":"integer"},"totalSilver":{"type":"integer","format":"int64"},"weightedAverageUnitPrice":{"type":"number","format":"double"}}},"NormalizedHistory":{"type":"object","properties":{"schemaVersion":{"type":"integer","const":1},"kind":{"type":"string","const":"market-history"},"source":{"type":"string"},"server":{"type":"string","enum":["west","east","europe"]},"item":{"$ref":"#/components/schemas/ItemDimension"},"location":{"$ref":"#/components/schemas/LocationDimension"},"quality":{"$ref":"#/components/schemas/QualityDimension"},"period":{"type":"string"},"capturedAt":{"type":"string","format":"date-time"},"summary":{"$ref":"#/components/schemas/NormalizedHistorySummary"},"history":{"type":"array","items":{"$ref":"#/components/schemas/NormalizedHistoryPoint"}},"dedupeKey":{"type":"string"}}},"NormalizedMarketOrder":{"type":"object","properties":{"schemaVersion":{"type":"integer","const":1},"kind":{"type":"string","const":"market-order"},"source":{"type":"string"},"server":{"type":"string","enum":["west","east","europe"]},"capturedAt":{"type":"string","format":"date-time"},"orderId":{"type":"integer","format":"int64"},"item":{"$ref":"#/components/schemas/ItemDimension"},"itemGroupId":{"type":"string"},"enchantmentLevel":{"type":"integer","minimum":0},"location":{"$ref":"#/components/schemas/LocationDimension"},"quality":{"$ref":"#/components/schemas/QualityDimension"},"auctionType":{"type":"string"},"side":{"type":"string","enum":["buy","sell"]},"unitPrice":{"type":"integer","format":"int64"},"amount":{"type":"integer","format":"int64"},"expiresAt":{"type":"string","format":"date-time"},"dedupeKey":{"type":"string"}}},"PriceRow":{"type":"object","properties":{"server":{"type":"string"},"itemIdentifier":{"type":"string"},"marketKey":{"type":"string"},"location":{"$ref":"#/components/schemas/LocationDimension"},"quality":{"type":"integer","minimum":1,"maximum":5},"sellPriceMin":{"anyOf":[{"type":"integer","format":"int64"},{"type":"null"}]},"sellPriceMinDate":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}]},"buyPriceMax":{"anyOf":[{"type":"integer","format":"int64"},{"type":"null"}]},"buyPriceMaxDate":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}]}}},"HistoryResponse":{"type":"object","properties":{"count":{"type":"integer","minimum":0},"data":{"type":"array","items":{"$ref":"#/components/schemas/NormalizedHistory"}},"source":{"type":"string","const":"local-market-service"}}},"OrdersResponse":{"type":"object","properties":{"count":{"type":"integer","minimum":0},"data":{"type":"array","items":{"$ref":"#/components/schemas/NormalizedMarketOrder"}},"source":{"type":"string","const":"local-market-service"}}},"PricesResponse":{"type":"object","properties":{"count":{"type":"integer","minimum":0},"data":{"type":"array","items":{"$ref":"#/components/schemas/PriceRow"}},"source":{"type":"string","const":"local-market-service"},"calculatedAt":{"type":"string","format":"date-time"}}},"MarketsResponse":{"type":"object","properties":{"count":{"type":"integer","minimum":0},"data":{"type":"array","items":{"$ref":"#/components/schemas/MarketDefinition"}},"source":{"type":"string","const":"local-market-service"}}},"RepositoryStats":{"type":"object","properties":{"historySnapshots":{"type":"integer","minimum":0},"orderSnapshots":{"type":"integer","minimum":0},"updatedAt":{"type":"string","format":"date-time"},"storage":{"type":"string"}}},"QueueSnapshot":{"type":"object","properties":{"depth":{"type":"integer"},"capacity":{"type":"integer"},"utilization_percent":{"type":"number"},"high_watermark":{"type":"integer"}}},"OutboxPipelineSnapshot":{"type":"object","properties":{"path":{"type":"string"},"pending_entries":{"type":"integer"},"pending_batches":{"type":"integer"},"retrying_batches":{"type":"integer"},"processing_batches":{"type":"integer"},"dead_letter_batches":{"type":"integer"},"oldest_pending_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}]},"oldest_pending_age_seconds":{"type":"integer","format":"int64"},"recovered_batches_total":{"type":"integer","minimum":0},"completed_batches_total":{"type":"integer","minimum":0},"rescheduled_batches_total":{"type":"integer","minimum":0},"dead_letter_batches_total":{"type":"integer","minimum":0}}},"TotalsSnapshot":{"type":"object","properties":{"enqueued_entries":{"type":"integer","minimum":0},"queue_dropped_entries":{"type":"integer","minimum":0},"batches_started":{"type":"integer","minimum":0},"batches_sent":{"type":"integer","minimum":0},"entries_sent":{"type":"integer","minimum":0},"send_attempts":{"type":"integer","minimum":0},"retries":{"type":"integer","minimum":0},"recovered_batches":{"type":"integer","minimum":0},"rescheduled_batches":{"type":"integer","minimum":0},"failed_batches":{"type":"integer","minimum":0},"dead_letter_batches":{"type":"integer","minimum":0},"entries_dropped_after_retries":{"type":"integer","minimum":0}}},"HistoryTotalsSnapshot":{"type":"object","properties":{"enqueued_entries":{"type":"integer","minimum":0},"enqueued_buckets":{"type":"integer","minimum":0},"queue_dropped_entries":{"type":"integer","minimum":0},"queue_dropped_buckets":{"type":"integer","minimum":0},"batches_started":{"type":"integer","minimum":0},"batches_sent":{"type":"integer","minimum":0},"entries_sent":{"type":"integer","minimum":0},"buckets_sent":{"type":"integer","minimum":0},"send_attempts":{"type":"integer","minimum":0},"retries":{"type":"integer","minimum":0},"recovered_batches":{"type":"integer","minimum":0},"rescheduled_batches":{"type":"integer","minimum":0},"failed_batches":{"type":"integer","minimum":0},"dead_letter_batches":{"type":"integer","minimum":0},"entries_dropped_after_retries":{"type":"integer","minimum":0},"buckets_dropped_after_retries":{"type":"integer","minimum":0}}},"LatencySnapshot":{"type":"object","properties":{"last_batch_ms":{"type":"number"},"average_batch_ms":{"type":"number"},"max_batch_ms":{"type":"number"},"last_attempt_ms":{"type":"number"}}},"ForwarderSnapshot":{"type":"object","properties":{"enabled":{"type":"boolean"},"running":{"type":"boolean"},"status":{"type":"string"},"in_flight_batches":{"type":"integer"},"queue":{"$ref":"#/components/schemas/QueueSnapshot"},"outbox":{"$ref":"#/components/schemas/OutboxPipelineSnapshot"},"totals":{"$ref":"#/components/schemas/TotalsSnapshot"},"latency_ms":{"$ref":"#/components/schemas/LatencySnapshot"},"last_status_code":{"type":"integer"},"last_success_at":{"type":"string","format":"date-time"},"last_error_at":{"type":"string","format":"date-time"},"last_error":{"type":"string"}}},"HistoryForwarderSnapshot":{"type":"object","properties":{"enabled":{"type":"boolean"},"running":{"type":"boolean"},"status":{"type":"string"},"in_flight_batches":{"type":"integer"},"queue":{"$ref":"#/components/schemas/QueueSnapshot"},"outbox":{"$ref":"#/components/schemas/OutboxPipelineSnapshot"},"totals":{"$ref":"#/components/schemas/HistoryTotalsSnapshot"},"latency_ms":{"$ref":"#/components/schemas/LatencySnapshot"},"last_status_code":{"type":"integer"},"last_success_at":{"type":"string","format":"date-time"},"last_error_at":{"type":"string","format":"date-time"},"last_error":{"type":"string"}}},"StatusResponse":{"type":"object","properties":{"status":{"type":"string","enum":["ok","degraded"]},"service":{"type":"string"},"environment":{"type":"string"},"source":{"type":"string","const":"local-market-service"},"uptime_seconds":{"type":"integer","format":"int64","minimum":0},"repository":{"$ref":"#/components/schemas/RepositoryStats"},"forwarder":{"$ref":"#/components/schemas/ForwarderSnapshot"},"price_forwarder":{"$ref":"#/components/schemas/ForwarderSnapshot"},"history_forwarder":{"$ref":"#/components/schemas/HistoryForwarderSnapshot"}}}}}}